@@ -2,7 +2,7 @@ require "spec_helper"
 
 describe Client::CampaignsController do
 
-  let(:account) { Factory(:account, :activated => true, api_key: "abc123") }
+  let(:account) { Factory(:account, :activated => true) }
   let(:user) { Factory(:user, :account => account) }
 
   describe "html format" do
@@ -71,7 +71,7 @@ describe Client::CampaignsController do
         preview_campaign = Factory(:preview, :account => account, :active => true)
         predictive_campaign = Factory(:predictive, :account => account, :active => true)
         inactive_campaign = Factory(:progressive, :account => account, :active => false)
-        get :index, :api_key=> 'abc123', :format => "json"
+        get :index, :api_key=> account.api_key, :format => "json"
         JSON.parse(response.body).length.should eq(2)
       end
 
@@ -88,7 +88,7 @@ describe Client::CampaignsController do
     describe "show" do
       it "should give campaign details" do
         campaign = Factory(:predictive, :account => account, :active => true, name: "Campaign 1")
-        get :show, :id=> campaign.id, :api_key=> 'abc123', :format => "json"
+        get :show, :id=> campaign.id, :api_key=> account.api_key, :format => "json"
         JSON.parse(response.body)['predictive']['name'].should eq("Campaign 1")
       end
     end
@@ -96,7 +96,7 @@ describe Client::CampaignsController do
     describe "edit" do
       it "should give campaign details" do
         predictive_campaign = Factory(:predictive, :account => account, :active => true, start_time: Time.now, end_time: Time.now, name: "Campaign 2")
-        get :edit, :id=> predictive_campaign.id, :api_key=> 'abc123', :format => "json"
+        get :edit, :id=> predictive_campaign.id, :api_key=> account.api_key, :format => "json"
         JSON.parse(response.body)['predictive']['name'].should eq("Campaign 2")
       end
     end
@@ -104,23 +104,23 @@ describe Client::CampaignsController do
     describe "destroy" do
       it "should delete campaign" do
         predictive_campaign = Factory(:predictive, :account => account, :active => true, start_time: Time.now, end_time: Time.now)
-        delete :destroy, :id=> predictive_campaign.id, :api_key=> 'abc123', :format => "json"
+        delete :destroy, :id=> predictive_campaign.id, :api_key=> account.api_key, :format => "json"
         response.body.should == "{\"message\":\"Campaign deleted\"}"
       end
 
       it "should not delete a campaign from another account" do
-        another_account = Factory(:account, :activated => true, api_key: "123abc")
+        another_account = Factory(:account, :activated => true)
         another_user = Factory(:user, account_id: another_account.id)
 
         predictive_campaign = Factory(:predictive, :account => account, :active => true, start_time: Time.now, end_time: Time.now)
-        delete :destroy, :id=> predictive_campaign.id, :api_key=> '123abc', :format => "json"
+        delete :destroy, :id=> predictive_campaign.id, :api_key=> another_account.api_key, :format => "json"
         response.body.should == "{\"message\":\"Cannot access campaign.\"}"
       end
 
       it "should not delete and return validation error" do
         caller = Factory(:caller)
         predictive_campaign = Factory(:predictive, :account => account, :active => true, start_time: Time.now, end_time: Time.now, callers: [caller])
-        delete :destroy, :id=> predictive_campaign.id, :api_key=> 'abc123', :format => "json"
+        delete :destroy, :id=> predictive_campaign.id, :api_key=> account.api_key, :format => "json"
         response.body.should == "{\"errors\":{\"caller_id\":[],\"base\":[\"There are currently callers assigned to this campaign. Please assign them to another campaign before deleting this one.\"]}}"
       end
     end
@@ -131,7 +131,7 @@ describe Client::CampaignsController do
         callers = 3.times.map{Factory(:caller, :account => account)}
         lambda {
           post :create , :campaign => {name: "abc", caller_id:"1234567890", script_id: script.id,
-            type: "Preview", time_zone: "Pacific Time (US & Canada)", start_time:  Time.new(2011, 1, 1, 9, 0, 0), end_time: Time.new(2011, 1, 1, 21, 0, 0)}, :api_key=> "abc123", :format => "json"
+            type: "Preview", time_zone: "Pacific Time (US & Canada)", start_time:  Time.new(2011, 1, 1, 9, 0, 0), end_time: Time.new(2011, 1, 1, 21, 0, 0)}, :api_key=> account.api_key, :format => "json"
         }.should change {account.reload.campaigns.size} .by(1)
         JSON.parse(response.body)['campaign']['name'].should  eq('abc')
       end
@@ -141,7 +141,7 @@ describe Client::CampaignsController do
         callers = 3.times.map{Factory(:caller, :account => account)}
         lambda {
           post :create , :campaign => {name: "abc", caller_id:"123456", script_id: script.id,
-            type: "Preview", time_zone: "Pacific Time (US & Canada)", start_time:  Time.new(2011, 1, 1, 9, 0, 0), end_time: Time.new(2011, 1, 1, 21, 0, 0)}, :api_key=> "abc123", :format => "json"
+            type: "Preview", time_zone: "Pacific Time (US & Canada)", start_time:  Time.new(2011, 1, 1, 9, 0, 0), end_time: Time.new(2011, 1, 1, 21, 0, 0)}, :api_key=> account.api_key, :format => "json"
         }.should change {account.reload.campaigns.size} .by(0)
         response.body.should eq("{\"errors\":{\"caller_id\":[],\"base\":[\"Caller ID must be a 10-digit North American phone number or begin with \\\"+\\\" and the country code\"]}}")
 
@@ -153,7 +153,7 @@ describe Client::CampaignsController do
       it "should update an existing campaign" do
         campaign = Factory(:predictive, name: "abc", account: account)
         lambda {
-          put :update , id: campaign.id, :campaign => {name: "def"}, :api_key=> "abc123", :format => "json"
+          put :update , id: campaign.id, :campaign => {name: "def"}, :api_key=> account.api_key, :format => "json"
         }.should change {account.reload.campaigns.size} .by(0)
         response.body.should  eq("{\"message\":\"Campaign updated\"}")
       end
@@ -162,7 +162,7 @@ describe Client::CampaignsController do
         voter_list = Factory(:voter_list, enabled: true)
         campaign = Factory(:predictive, name: "abc", account: account, voter_lists: [voter_list])
         lambda {
-          put :update , id: campaign.id, :campaign => {name: "def", :voter_lists_attributes=> {"0"=>{"id"=>"#{voter_list.id}", "enabled"=> "0"}}}, :api_key=> "abc123", :format => "json"
+          put :update , id: campaign.id, :campaign => {name: "def", :voter_lists_attributes=> {"0"=>{"id"=>"#{voter_list.id}", "enabled"=> "0"}}}, :api_key=> account.api_key, :format => "json"
         }.should change {account.reload.campaigns.size} .by(0)
         response.body.should  eq("{\"message\":\"Campaign updated\"}")
         voter_list.reload.enabled.should be_false
@@ -172,7 +172,7 @@ describe Client::CampaignsController do
       it "should throw validation error" do
         campaign = Factory(:predictive, name: "abc", account: account)
         lambda {
-          put :update , id: campaign.id, :campaign => {caller_id: "123"}, :api_key=> "abc123", :format => "json"
+          put :update , id: campaign.id, :campaign => {caller_id: "123"}, :api_key=> account.api_key, :format => "json"
         }.should change {account.reload.campaigns.size} .by(0)
         response.body.should  eq("{\"errors\":{\"caller_id\":[],\"base\":[\"Caller ID must be a 10-digit North American phone number or begin with \\\"+\\\" and the country code\"]}}")
       end
@@ -185,7 +185,7 @@ describe Client::CampaignsController do
       it "should show deleted campaigns" do
         manual_campaign = Factory(:preview, :account => account, :active => true)
         inactive_campaign = Factory(:progressive, :account => account, :active => false)
-        get :deleted, :api_key=> 'abc123', :format => "json"
+        get :deleted, :api_key=> account.api_key, :format => "json"
         JSON.parse(response.body).length.should eq(1)
       end
     end
@@ -194,7 +194,7 @@ describe Client::CampaignsController do
 
       it "should restore inactive campaign" do
         inactive_campaign = Factory(:progressive, :account => account, :active => false)
-        put :restore, campaign_id: inactive_campaign.id, :api_key=> 'abc123', :format => "json"
+        put :restore, campaign_id: inactive_campaign.id, :api_key=> account.api_key, :format => "json"
         response.body.should eq("{\"message\":\"Campaign restored\"}")
       end
     end
