@@ -1,7 +1,7 @@
 require "spec_helper"
 
 describe Client::PossibleResponsesController do
-  let(:account) { Factory(:account, api_key: "abc123") }
+  let(:account) { Factory(:account) }
   before(:each) do
     @user = Factory(:user, account_id: account.id)
   end
@@ -14,7 +14,7 @@ describe Client::PossibleResponsesController do
       question = Factory(:question, :script => active_script)
       possible_response1 = Factory(:possible_response, question_id: question.id)
       possible_response2 = Factory(:possible_response, question_id: question.id)
-      get :index, script_id: active_script.id, question_id: question.id, :api_key=> 'abc123', :format => "json"
+      get :index, script_id: active_script.id, question_id: question.id, :api_key=> account.api_key, :format => "json"
       response.body.should eq("[#{possible_response1.to_json},#{possible_response2.to_json}]")
     end
   end
@@ -24,7 +24,7 @@ describe Client::PossibleResponsesController do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, :script => active_script)
       possible_response = Factory(:possible_response, question_id: question.id)
-      get :show, script_id: active_script.id, question_id: question.id, id: possible_response.id, :api_key=> 'abc123', :format => "json"
+      get :show, script_id: active_script.id, question_id: question.id, id: possible_response.id, :api_key=> account.api_key, :format => "json"
       response.body.should eq("{\"possible_response\":{\"external_id_field\":null,\"id\":#{possible_response.id},\"keypad\":null,\"possible_response_order\":1,\"question_id\":#{question.id},\"retry\":false,\"value\":\"no_response\"}}")
     end
 
@@ -32,7 +32,7 @@ describe Client::PossibleResponsesController do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, :script => active_script)
       possible_response = Factory(:possible_response, question_id: question.id)
-      get :show, script_id: 100, question_id: question.id, id: possible_response.id, :api_key=> 'abc123', :format => "json"
+      get :show, script_id: 100, question_id: question.id, id: possible_response.id, :api_key=> account.api_key, :format => "json"
       response.body.should eq("{\"message\":\"Resource not found\"}")
     end
 
@@ -40,7 +40,7 @@ describe Client::PossibleResponsesController do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, :script => active_script)
       possible_response = Factory(:possible_response, question_id: question.id)
-      get :show, script_id: active_script.id, question_id: 100, id: possible_response.id, :api_key=> 'abc123', :format => "json"
+      get :show, script_id: active_script.id, question_id: 100, id: possible_response.id, :api_key=> account.api_key, :format => "json"
       response.body.should eq("{\"message\":\"Resource not found\"}")
     end
 
@@ -48,7 +48,7 @@ describe Client::PossibleResponsesController do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, :script => active_script)
       possible_response = Factory(:possible_response, question_id: question.id)
-      get :show, script_id: active_script.id, question_id: question.id, id: 100, :api_key=> 'abc123', :format => "json"
+      get :show, script_id: active_script.id, question_id: question.id, id: 100, :api_key=> account.api_key, :format => "json"
       response.body.should eq("{\"message\":\"Resource not found\"}")
     end
 
@@ -59,7 +59,7 @@ describe Client::PossibleResponsesController do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, text: "abc", script_order: 1, script: active_script)
       possible_response = Factory(:possible_response, question_id: question.id)
-      delete :destroy, script_id: active_script.id, question_id: question.id, id: possible_response.id,  :api_key=> 'abc123', :format => "json"
+      delete :destroy, script_id: active_script.id, question_id: question.id, id: possible_response.id,  :api_key=> account.api_key, :format => "json"
       response.body.should eq("{\"message\":\"Possible Response Deleted\",\"status\":\"ok\"}")
     end
   end
@@ -68,14 +68,14 @@ describe Client::PossibleResponsesController do
     it "should create possible response" do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, text: "abc", script_order: 1, script: active_script)
-      post :create, script_id: active_script.id, question_id: question.id, possible_response: {value: "Hi", possible_response_order: 1},  :api_key=> 'abc123', :format => "json"
+      post :create, script_id: active_script.id, question_id: question.id, possible_response: {value: "Hi", possible_response_order: 1},  :api_key=> account.api_key, :format => "json"
       response.body.should match(/{\"possible_response\":{\"external_id_field\":null,\"id\":(.*),\"keypad\":null,\"possible_response_order\":1,\"question_id\":#{question.id},\"retry\":false,\"value\":\"Hi\"}}/)
     end
 
     it "should throw validation error" do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, text: "abc", script_order: 1, script: active_script)
-      post :create, script_id: active_script.id, question_id: question.id, possible_response: {value: nil, possible_response_order: 1},  :api_key=> 'abc123', :format => "json"
+      post :create, script_id: active_script.id, question_id: question.id, possible_response: {value: nil, possible_response_order: 1},  :api_key=> account.api_key, :format => "json"
       response.body.should eq("{\"errors\":{\"value\":[\"can't be blank\"]}}")
     end
 
@@ -86,7 +86,7 @@ describe Client::PossibleResponsesController do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, text: "abc", script_order: 1, script: active_script)
       possible_response = Factory(:possible_response, question_id: question.id)
-      put :update, script_id: active_script.id, question_id: question.id, id: possible_response.id, possible_response: {value: "Hi"},  :api_key=> 'abc123', :format => "json"
+      put :update, script_id: active_script.id, question_id: question.id, id: possible_response.id, possible_response: {value: "Hi"},  :api_key=> account.api_key, :format => "json"
       response.body.should eq("{\"message\":\"Possible Response updated\"}")
       possible_response.reload.value.should eq("Hi")
     end
@@ -95,7 +95,7 @@ describe Client::PossibleResponsesController do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, text: "abc", script_order: 1, script: active_script)
       possible_response = Factory(:possible_response, question_id: question.id)
-      put :update, script_id: active_script.id, question_id: question.id, id: possible_response.id, possible_response: {value: nil},  :api_key=> 'abc123', :format => "json"
+      put :update, script_id: active_script.id, question_id: question.id, id: possible_response.id, possible_response: {value: nil},  :api_key=> account.api_key, :format => "json"
       response.body.should eq("{\"errors\":{\"value\":[\"can't be blank\"]}}")
     end
 
