@@ -1,7 +1,7 @@
 require "spec_helper"
 
 describe Client::QuestionsController do
-  let(:account) { Factory(:account, api_key: "abc123") }
+  let(:account) { Factory(:account) }
   before(:each) do
     @user = Factory(:user, account_id: account.id)
   end
@@ -13,7 +13,7 @@ describe Client::QuestionsController do
       active_script = Factory(:script, :account => account, :active => true)
       question1 = Factory(:question, text: "abc", script_order: 1, script: active_script)
       question2 = Factory(:question, text: "def", script_order: 2, script: active_script)
-      get :index, script_id: active_script.id, :api_key=> 'abc123', :format => "json"
+      get :index, script_id: active_script.id, :api_key=> account.api_key, :format => "json"
       response.body.should eq("[#{question1.to_json},#{question2.to_json}]")
     end
   end
@@ -22,7 +22,7 @@ describe Client::QuestionsController do
     it "should return question " do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, text: "abc", script_order: 1, script: active_script)
-      get :show, script_id: active_script.id, id: question.id,  :api_key=> 'abc123', :format => "json"
+      get :show, script_id: active_script.id, id: question.id,  :api_key=> account.api_key, :format => "json"
       response.body.should eq "{\"id\":#{question.id},\"text\":\"abc\",\"script_order\":1,\"external_id_field\":null,\"script_id\":#{active_script.id},\"possible_responses\":[]}"
 
     end
@@ -30,14 +30,14 @@ describe Client::QuestionsController do
     it "should 404 if script not found" do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, text: "abc", script_order: 1, script: active_script)
-      get :show, script_id: 100, id: question.id,  :api_key=> 'abc123', :format => "json"
+      get :show, script_id: 100, id: question.id,  :api_key=> account.api_key, :format => "json"
       response.body.should eq("{\"message\":\"Resource not found\"}")
     end
 
     it "should 404 if question not found in script" do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, text: "abc", script_order: 1, script: active_script)
-      get :show, script_id: active_script.id, id: 100,  :api_key=> 'abc123', :format => "json"
+      get :show, script_id: active_script.id, id: 100,  :api_key=> account.api_key, :format => "json"
       response.body.should eq("{\"message\":\"Resource not found\"}")
     end
 
@@ -48,7 +48,7 @@ describe Client::QuestionsController do
     it "should delete question" do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, text: "abc", script_order: 1, script: active_script)
-      delete :destroy, script_id: active_script.id, id: question.id,  :api_key=> 'abc123', :format => "json"
+      delete :destroy, script_id: active_script.id, id: question.id,  :api_key=> account.api_key, :format => "json"
       response.body.should eq("{\"message\":\"Question Deleted\",\"status\":\"ok\"}")
     end
   end
@@ -56,13 +56,13 @@ describe Client::QuestionsController do
   describe "create" do
     it "should create question" do
       active_script = Factory(:script, :account => account, :active => true)
-      post :create, script_id: active_script.id, question: {text: "Hi", script_order: 1},  :api_key=> 'abc123', :format => "json"
+      post :create, script_id: active_script.id, question: {text: "Hi", script_order: 1},  :api_key=> account.api_key, :format => "json"
       response.body.should eq "{\"id\":#{active_script.questions.first.id},\"text\":\"Hi\",\"script_order\":1,\"external_id_field\":null,\"script_id\":#{active_script.id},\"possible_responses\":[]}"
     end
 
     it "should throw validation error" do
       active_script = Factory(:script, :account => account, :active => true)
-      post :create, script_id: active_script.id, question: {text: "Hi"},  :api_key=> 'abc123', :format => "json"
+      post :create, script_id: active_script.id, question: {text: "Hi"},  :api_key=> account.api_key, :format => "json"
       response.body.should eq("{\"errors\":{\"script_order\":[\"can't be blank\",\"is not a number\"]}}")
     end
 
@@ -72,7 +72,7 @@ describe Client::QuestionsController do
     it "should update a question" do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, text: "abc", script_order: 1, script: active_script)
-      put :update, script_id: active_script.id, id: question.id, question: {text: "Hi"},  :api_key=> 'abc123', :format => "json"
+      put :update, script_id: active_script.id, id: question.id, question: {text: "Hi"},  :api_key=> account.api_key, :format => "json"
       response.body.should eq("{\"message\":\"Question updated\"}")
       question.reload.text.should eq("Hi")
     end
@@ -80,7 +80,7 @@ describe Client::QuestionsController do
     it "should throw validation error" do
       active_script = Factory(:script, :account => account, :active => true)
       question = Factory(:question, text: "abc", script_order: 1, script: active_script)
-      put :update, script_id: active_script.id, id: question.id, question: {text: "Hi", script_order: nil},  :api_key=> 'abc123', :format => "json"
+      put :update, script_id: active_script.id, id: question.id, question: {text: "Hi", script_order: nil},  :api_key=> account.api_key, :format => "json"
       response.body.should eq("{\"errors\":{\"script_order\":[\"can't be blank\",\"is not a number\"]}}")
     end
 
