@@ -3,7 +3,7 @@ require "spec_helper"
 describe VoterListsController do
 
   describe "api" do
-    let(:account) { Factory(:account, api_key: "abc123") }
+    let(:account) { Factory(:account) }
     before(:each) do
       @user = Factory(:user, account_id: account.id)
     end
@@ -12,7 +12,7 @@ describe VoterListsController do
       it "should list voter lists for a campaign" do
         voter_list = Factory(:voter_list)
         campaign = Factory(:campaign, :account => account, :active => true, voter_lists: [voter_list])
-        get :index, campaign_id: campaign.id, :api_key=> 'abc123', :format => "json"
+        get :index, campaign_id: campaign.id, :api_key=> account.api_key, :format => "json"
         response.body.should eq( "[{\"voter_list\":{\"enabled\":true,\"id\":#{voter_list.id},\"name\":\"#{voter_list.name}\"}}]")
       end
     end
@@ -21,21 +21,21 @@ describe VoterListsController do
       it "should shows voter list " do
         voter_list = Factory(:voter_list)
         campaign = Factory(:campaign, :account => account, :active => true, voter_lists: [voter_list, Factory(:voter_list)])
-        get :show, campaign_id: campaign.id, id: voter_list.id, :api_key=> 'abc123', :format => "json"
+        get :show, campaign_id: campaign.id, id: voter_list.id, :api_key=> account.api_key, :format => "json"
         response.body.should eq( "{\"voter_list\":{\"enabled\":true,\"id\":#{voter_list.id},\"name\":\"#{voter_list.name}\"}}")
       end
 
       it "should throws 404 if campaign not found " do
         voter_list = Factory(:voter_list)
         campaign = Factory(:campaign, :account => account, :active => true, voter_lists: [voter_list, Factory(:voter_list)])
-        get :show, campaign_id: 100, id: voter_list.id, :api_key=> 'abc123', :format => "json"
+        get :show, campaign_id: 100, id: voter_list.id, :api_key=> account.api_key, :format => "json"
         response.body.should eq( "{\"message\":\"Resource not found\"}" )
       end
 
       it "should throws 404 if voter list not found " do
         voter_list = Factory(:voter_list)
         campaign = Factory(:campaign, :account => account, :active => true, voter_lists: [voter_list, Factory(:voter_list)])
-        get :show, campaign_id: campaign.id, id: 100, :api_key=> 'abc123', :format => "json"
+        get :show, campaign_id: campaign.id, id: 100, :api_key=> account.api_key, :format => "json"
         response.body.should eq( "{\"message\":\"Resource not found\"}" )
       end
 
@@ -45,7 +45,7 @@ describe VoterListsController do
       it "should enable voter list " do
         voter_list = Factory(:voter_list, enabled: false)
         campaign = Factory(:campaign, :account => account, :active => true, voter_lists: [voter_list, Factory(:voter_list)])
-        put :enable, campaign_id: campaign.id, id: voter_list.id, :api_key=> 'abc123', :format => "json"
+        put :enable, campaign_id: campaign.id, id: voter_list.id, :api_key=> account.api_key, :format => "json"
         response.body.should eq( "{\"message\":\"Voter List enabled\"}")
       end
 
@@ -55,7 +55,7 @@ describe VoterListsController do
       it "should disable voter list " do
         voter_list = Factory(:voter_list, enabled: true)
         campaign = Factory(:campaign, :account => account, :active => true, voter_lists: [voter_list, Factory(:voter_list)])
-        put :disable, campaign_id: campaign.id, id: voter_list.id, :api_key=> 'abc123', :format => "json"
+        put :disable, campaign_id: campaign.id, id: voter_list.id, :api_key=> account.api_key, :format => "json"
         response.body.should eq( "{\"message\":\"Voter List disabled\"}")
       end
 
@@ -65,7 +65,7 @@ describe VoterListsController do
       it "should update voter list " do
         voter_list = Factory(:voter_list, enabled: true, name: "abc")
         campaign = Factory(:campaign, :account => account, :active => true, voter_lists: [voter_list, Factory(:voter_list)])
-        put :update, campaign_id: campaign.id, id: voter_list.id, voter_list: {name: "xyz"}, :api_key=> 'abc123', :format => "json"
+        put :update, campaign_id: campaign.id, id: voter_list.id, voter_list: {name: "xyz"}, :api_key=> account.api_key, :format => "json"
         response.body.should eq( "{\"message\":\"Voter List updated\"}" )
         voter_list.reload.name.should  eq('xyz')
       end
@@ -75,7 +75,7 @@ describe VoterListsController do
       it "should update voter list " do
         voter_list = Factory(:voter_list, enabled: true, name: "abc")
         campaign = Factory(:campaign, :account => account, :active => true, voter_lists: [voter_list, Factory(:voter_list)])
-        delete :destroy, campaign_id: campaign.id, id: voter_list.id, :api_key=> 'abc123', :format => "json"
+        delete :destroy, campaign_id: campaign.id, id: voter_list.id, :api_key=> account.api_key, :format => "json"
         response.body.should eq( "{\"message\":\"This opeartion is not permitted\"}")
       end
     end
@@ -87,7 +87,7 @@ describe VoterListsController do
         campaign = Factory(:campaign, :account => account, :active => true)
         Resque.should_receive(:enqueue)
         post :create, campaign_id: campaign.id, voter_list: {name: "abc.csv", separator: ",", headers: "[]", csv_to_system_map: "{\"Phone\": \"Phone\"}",
-        s3path: "abc"}, upload: csv_file_upload, :api_key=> 'abc123', :format => "json"
+        s3path: "abc"}, upload: csv_file_upload, :api_key=> account.api_key, :format => "json"
         response.body.should match(/{\"voter_list\":{\"enabled\":true,\"id\":(.*),\"name\":\"abc.csv\"}}/)
       end
 
@@ -95,14 +95,14 @@ describe VoterListsController do
         csv_file_upload =  {"datafile" => fixture_file_upload("/files/voter_list.xsl")}
         campaign = Factory(:campaign, :account => account, :active => true)
         post :create, campaign_id: campaign.id, voter_list: {name: "abc", separator: ",", headers: "[]", csv_to_system_map: "{\"Phone\": \"Phone\"}",
-        s3path: "abc"}, upload: csv_file_upload, :api_key=> 'abc123', :format => "json"
+        s3path: "abc"}, upload: csv_file_upload, :api_key=> account.api_key, :format => "json"
         response.body.should eq("{\"errors\":{\"base\":[\"Wrong file format. Please upload a comma-separated value (CSV) or tab-delimited text (TXT) file. If your list is in Excel format (XLS or XLSX), use \\\"Save As\\\" to change it to one of these formats.\"]}}")
       end
 
       it "should throw validation error if file not uploaded" do
         campaign = Factory(:campaign, :account => account, :active => true)
         post :create, campaign_id: campaign.id, voter_list: {name: "abc", separator: ",", headers: "[]", csv_to_system_map: "{\"Phone\": \"Phone\"}",
-        s3path: "abc"}, upload: nil, :api_key=> 'abc123', :format => "json"
+        s3path: "abc"}, upload: nil, :api_key=> account.api_key, :format => "json"
         response.body.should eq("{\"errors\":{\"uploaded_file_name\":[\"can't be blank\"],\"base\":[\"Please upload a file.\"]}}")
       end
 
