@@ -1,5 +1,5 @@
 module WhiteLabeling
-  ['title', 'full_title', 'phone', 'email','billing_link' ].each do |value|
+  ['title', 'full_title', 'phone', 'email','billing_link', "add_to_balance_link", "configure_auto_recharge_link" ].each do |value|
     define_method("white_labeled_#{value}") do |domain|
       I18n.t("white_labeling.#{correct_domain(domain)}.#{value}")
     end
@@ -14,5 +14,5 @@ module WhiteLabeling
       d
     end
   end
-  
+
 end
