@@ -23,7 +23,7 @@ ImpactDialing::Application.configure do
   # The production environment is meant for finished, "live" apps.
   # Code is not reloaded between requests
   config.cache_classes = true
-  config.lograge.enabled = true
+  config.lograge.enabled = false
 
   # Full error reports are disabled and caching is turned on
   config.consider_all_requests_local = false
